feat: basic JS API
